Allow specifying a custom profile for a JRE mapping.
Currently `ant4eclipe` tries to determine the runtime profile from the
installed JRE. This is fine, if the installed JRE matches the profile
one wants to build with. In both of the following cases, this is
sub-optimal:

 * The installed JRE natively only supports a newer profile than one
wants to build against.
 * The profile of the installed JRE is not known by `ant4eclipse`.

In both cases, it would be useful, if one could specify the profile to
use for a configured JRE:

```
<jre id="JavaSE-1.8" location="${java.home}" profile="JavaSE-1.8"/>
```

This means that for the JRE identified by the ID `JavaSE-1.8` the JRE
should be used that is the default one (located at `java.home`) and
`ant4eclipe` should assume that it supports the profile `JavaSE-1.8` no
matter how the installed JRE does describe itself.

With this setting, it is possible to even build against a `JavaSE-1.8`
profile, if only a Java 11 JDK is installed.